5th Part Doubt
Had this doubt above too.
The button, which is an item of a flex row, stretches vertically.
My question is that why, why doesn't this same thing happen when I use flex columns, where I have to stretch the items.
Why when I put flex-grow as 0, that doesn't solve the problem either.